Add Call, CallFrame and MemoryRange
The memory range tuples in the VM are expected to behave as start/size
pairs. However, Rust uses start/end patterns. The `MemoryRange` type is
an abstraction to interface with the values provided by the bytecode
operations and Rust memory scheme.

The `Call` and `CallFrame` structures will heavily depend on that
because they have internal references to owned memory regions.