Update opcodes list (#9)
* Update opcodes list

The opcodes list was outdated and needed to be resync with the specs.

https://github.com/FuelLabs/fuel-specs/blob/master/specs/vm/opcodes.md

Also, the previous implementation didn't consider the different sizes of
immediate values, treating everything as u16.

The API still need to improve to prevent users from allocating invalid
register values with overflow.

Resolves #8

* Update interpreter to latest specs

The interpreter is not fully sync with the specs.

Also, the functionality scope can be split into different modules to
facilitate maintenance of the routines.

* Implement interpreter mem and crypto functions

Some of the functions contains particularities such as the ecrecover
that uses a special encoding regarding the recovery id. These are not
native to the public libraries and must be manually handled.

* Add test case for keccak256

This new test case is symmetric to the sha256 test case with the crypto
primitive as difference.

* Add Transaction structure

The transaction will define the VM initialization as well as the
underlying input/output/contract types.

* Implement recoverable secp256k1 signature

The EcRecover scheme may benefit from a compression that can save an
additional byte of storage.

This crypto primitive is implemented using rusk-secp256k1 as backend.

* Add MemClearI operation

* Rename 32-byte storage ops to "quad word"

* Rename opcodes to fit in 4 chars

* Add serialization for all tx types

* Change VM init to copy tx to stack

The initialization of the VM should copy the hash of the transaction,
its size and the serialized bytes to the stack area so the execution of
the program can access it when required.

* Update IO according to specs

The specs points that every number should be stored as u64 and the
arrays of bytes should be padded to multiples of 8.

* Change `crypto::hash` to encapsulate primitive

The underlying default hash function of the crypto module should not
impact the API and any change in the primitive should be transparent to
the users of the interpreter.

* Add Log opcode

This opcode will perform a naive dump of the selected registers ignoring
the zeroed indices.

This is a primitive implementation to a more complex log/event/topic
tracing system.

* Add binary interpreter

This binary executes transactions and output the log opcodes.

* Add tracing to interpreter execution

* Add debug to registers

The mutated registers should be print if the log level is set to debug.

* Update memcopy to non-overlapping

THe `MCP` operation should fail if the memory segments are overlapping.

This will allow the internals to benefit from `memcopy` instead of
`memmove`.

* Add `is_valid` function for all transaction types

* Add Tx validation to VM initialization

The VM initialization must validate the transaction before the registers
and memory are prepared for execution.

The TX validation is critical and most of the edge cases must have test
coverage.

* Optimize code and remove redundant checks

* Add Call, CallFrame and MemoryRange

The memory range tuples in the VM are expected to behave as start/size
pairs. However, Rust uses start/end patterns. The `MemoryRange` type is
an abstraction to interface with the values provided by the bytecode
operations and Rust memory scheme.

The `Call` and `CallFrame` structures will heavily depend on that
because they have internal references to owned memory regions.

* Add Call opcode implementation

The call frames structure and the reflection of a call in the registers
and memory is critical to the system and must be broadly tested.

This commit introduces the update of the program counter.

* Update dependencies to use fuel-* libs

The transaction and opcode structs were split into their own libraries.

This commit introduces the usage of these libraries as dependencies.

* Update ECR to clear err

In case of failure, ECRecover should set the recovered public-key to
`0` and turn the `err` flag on.

In case of success, `$rA` will be set with the recovered public-key and
the error flag will be cleared.

* Add predicate verification for tx create coin

* Add predicate verification

Predicate verification should initialize the VM in a special state and
execute the opcodes under special constraints defined in the specs.