Created a new entity - `banknote`. It is an aggregator for the coins and messages. We can't use `Input` or `Output` naming because we already do that with transactions. So I introduced a new one. We can think about the naming in the PR=)
Renamed `coinsToSpent` into `banknotesToSpent`. Move it to a separate file.
Changed the signature. Not it returns `Vec<Vec<Banknote>>` instead of `Vec<Coin>`.
The result is a list of `Banknote` per asset(we can have many different assets, so it is why it has an additional vector).
Replaced exclude vector with two vectors in the separate struct. One vector allows filtering by utxos, another by messages.

Reduced the number of operations in the `banknotesToSpent`.
Added some abstraction to remove logic duplicating.
I want to reuse it in the other schemas like `balance`.

Added `schema::Scalar` for `MessageId`.

I need to add the support of `Message` into tests and the code. I will do it in the next commit=)