Add support of messages into `coinsToSpend` (#591)
* Created a new entity - `banknote`. It is an aggregator for the coins and messages.
We can't use `Input` or `Output` naming because we already do that with transactions.
So I introduced a new one. We can think about the naming in the PR=)

Renamed `coinsToSpent` into `banknotesToSpent`. Move it to a separate file.
Changed the signature. Not it returns `Vec<Vec<Banknote>>` instead of `Vec<Coin>`.
The result is a list of `Banknote` per asset(we can have many different assets, so it is why it has an additional vector).
Replaced exclude vector with two vectors in the separate struct. One vector allows filtering by utxos, another by messages.

Reduced the number of operations in the `banknotesToSpent`.
Added some abstraction to remove logic duplicating.
I want to reuse it in the other schemas like `balance`.

Added `schema::Scalar` for `MessageId`.

I need to add the support of `Message` into tests and the code. I will do it in the next commit=)

* Merged with master

* Applied changes from the issue

* Added messages into the `unspent_resources` result.
Uncommented and fixed old tests.

* Make clippy happy)

* Make clippy happy)

* Added unit tests for `Message`. I used old test-cases and adapted them to work with coins, messages, and coins and messages test cases.
Added new field to `schema::Message` - `message_id`. It is calculated based on the `Message`.
Added check of errors in the client's tests.

* Updated teh comment

* Increase timeout to 45, if it helps, I will create a separate PR

* Remove whitespace

* Applied suggestions from teh review.

* Added link to the issue for each added "TODO".
Renamed `CoinQueryError` into `ResourceQueryError`.
Added `asset_id` and `collected_amount` into `ResourceQueryError`.

* Merge with storage refactoring