Add `fuel-benches` (#616)
This commit introduces the benchmarks for fuel-vm. It covers the first
phase of #256 and targets individual instructions with the
rocksdb/database backend.

This targets gas instrumentalization, so we skip benchmarking for
predicate cases.

This uses a dedicated component of fuel-core so it will be easier to
isolate and extend this component into block checking, and maybe to add
some script to parse the output from criterion.

The benchmark is created from an isolated state of the VM that is ready
to execute a given instruction. It is sub-optimal in the case of
contract state & call since the call overhead will always be added to
any instruction to be checked inside of a contract (i.e. SWW).

As next step, we need to make an API in fuel-vm to allow users to
perform the operations of a call without executing the contract code. In
other words, just load the contract code, add the call frame to the
stack and prepare the registers.

After this phase is complete, we need to benchmark beyond instructions
to a block level so the post-execution steps (database commit, etc) is
also accounted.