revert optimistic reserved node reconnect (#995)
During #974 optimistic reconnect was removed from the code as being a
_duplicate_ work.

It was not a duplicate work, more of an optimistic immediate reconnect
if the node is available, while `health_check` tries to reconnect after
a certain Interval (e.g. 30-60 seconds) which is there as a backup only
if the optimistic reconnect was not successful.

I also added a small description comment to `OutboundState`.