Graceful Shutdown for all services (#1018)
- Closes https://github.com/FuelLabs/fuel-core/issues/949
- Closes https://github.com/FuelLabs/fuel-core/issues/897

Removed second spawned task for `fuel_core_services::service::run`
function.

Instead, wrapped it into `catch_unwind`.
Added a `shutdown` method for the task that allows doing some cleanup
for methods. For, almost, all services, it does nothing, but can be
reused in the future. Added handling of the shutdown for `Relayer` to
act immediately instead of waiting for the synchronization's end.
Improved handling of the stop signal for `fuel-core-sync`.