feat: Use ID types in client api (#1191)
Related issues:
- Closes #1121

This PR update the `FuelClient` API to use ID types in place of string
serialized IDs. Now, consumers of the API must directly pass a client
type when performing a method call. Internally, these ID types are then
converted to GraphQL types to be used in the GraphQL API. This API
update represents a breaking change.