Remove `DiscoveryEvent` and use `KademliaEvent` instead (#1377)
Closes https://github.com/FuelLabs/fuel-core/issues/1065

The idea being that we had the `DiscoveryEvent` that didn't serve any
special purpose and wasn't being consumed anywhere. This simplifies
`DiscoveryBehavior`.

The issue also suggested:
> Upon removing the Discovery test should be updated too.

I'm not really sure what this test is for. The actual code uses `mdns`
to identify new peer addresses, whereas the test bypasses this and does
it manually on `UnroutablePeer` events. I don't understand the point of
this.