Moved insertion of the blocks into the `BlockImporter` instead of the executor (#1577)
Related to https://github.com/FuelLabs/fuel-core/issues/1549

We don't need to insert the block in the executor because it is the
block importer's responsibility. It verifies the block's validity and
decides whether we want to insert a new block or not. Plus, storing
blocks and transactions is not part of the state transition.

This change also adds the ability to produce the block with a defined
order of the transactions. It may be useful in the tests.