Remove nested contract resolvers in GQL schema (#1675)
closes: #1652

Having the ability to fetch all data related to a contract is overkill
and costly in most cases, and usually done by accident. This api change
simplifies the GQL schema to simply return a contractId instead of a
contract resolver for tx inputs, receipts, and more.

Note that the full contract details (such as the contract bytecode) can
still be fetched directly via the contract query endpoint.