Soften txpool p2p reputation requirements (#1663)
It is fairly common for gossipped transactions to fail due to transient
reasons such as UTXO's being spent before the transaction is included. A
transaction that was valid when broadcast could become invalid if its
inputs were spent in a newly confirmed block. Penalizing nodes for this
could unfairly punish nodes that are simply "behind" in their view of
the blockchain state.

In bitcoin, only incorrectly formatted transactions (i.e. invalid
signatures or corrupt data according to consensus rules) are considered
immediately punishable, and all transient (ie. state dependent) failures
are ignored but unpunished.
https://github.com/bitcoin/bitcoin/blob/6ff0aa089c01ff3e610ecb47814ed739d685a14c/src/net_processing.cpp#L1849

However, bitcoin does implement rate limiting on peers to prevent spam
of either valid or invalid transactions to work around this. That is not
currently implemented in this PR (but noted in this issue: #1677)

https://github.com/bitcoin/bitcoin/blob/6ff0aa089c01ff3e610ecb47814ed739d685a14c/src/net_processing.cpp#L3842C23-L3842C37