Extraction of the `FuelBlockSecondaryKeyBlockHeights` table to an off-chain worker (#1671)
Closes https://github.com/FuelLabs/fuel-core/issues/1583

The change moves the `FuelBlockSecondaryKeyBlockHeights` table and
related logic to the off-chain worker. Along with this, the change adds
a new `Merklelized` blueprint that maintains the binary Merkle tree over
the storage data. It supports only the insertion of the objects without
removing them. This blueprint replaces the logic that previously was
defined in the `Database`.

Some side changes caused by the main change:
- Now, each blueprint provides its own set of tests related to the logic
of this blueprint.
- The `TransactionStatus` uses `block_height` inside instead of the
`block_id`.
- The key for the dense merkle tree looks like:
  ```rust
  pub enum DenseMetadataKey<PrimaryKey> {
      /// The primary key of the `DenseMerkleMetadata`.
      Primary(PrimaryKey),
      #[default]
      /// The latest `DenseMerkleMetadata` of the table.
      Latest,
  }
  ```

---------

Co-authored-by: Voxelot <brandonkite92@gmail.com>