Keep record of events from L1 in Block Header (#1769)
closes https://github.com/FuelLabs/fuel-core/issues/1741

This PR introduces a new field on the block header generated section
called `event_inbox_root`. This is calculated after execution by adding
all the hashes of the events from the L1 relayer to a merkle tree and
generating the root.

Also renamed the message receipt root to `message_outbox_root` to keep
up the theme.