refactor: Prepare (re)genesis for off chain tables (#1771)
Mostly refactoring to allow for #1619 . Not final but close. Some
cleanup remains here or there. That will be bundled with the actual
feature itself.

The `SnapshotReader` and `SnapshotWriter` (previously the `StateReader`
and `StateWriter`) now:
* have a single `read` (or `write`) method generic over the table type
* no longer use `*Config` structs in the interface -- all reading and
writing is to be done over tables
* `ChainConfig` is now written and read by the `SnapshotWriter` and
`SnapshotReader` (since it is actually part of the snapshot).

The genesis progress is now a generic `String` -> `u64` mapping with no
enumerations for the key (previously we had an enum with a variant for
each table).

Every table gets its own parquet file. Json is still in a single file
with the `StateConfig` schema. Depending on whether we want the offchain
tables in the `StateConfig` we can either drop them or include them in
the next PR.

I avoided enumerating the tables as much as possible to lessen coupling.
A new table should ideally require:
1. a call to write::<NewTable> when generating the snapshot
2. an implementation of `ProcessState<NewTable>` to describe how to
import it on regenesis
3. an implementation of `AsTable` to describe how (if at all) this table
can be extracted from a in-memory `StateConfig`
4. a call to `workers.spawn::<NewTable>` to run a worker to import it
from the snapshot.

After this PR we'll add in the off-chain tables and identify and
regenerate dependent tables (ideally with batching + resumability).

---------

Co-authored-by: xgreenx <xgreenx9999@gmail.com>