Separate production from dry runs in executor & Cleanup all execution paths :) (#1873)
Closes: https://github.com/FuelLabs/fuel-core/issues/1751

- Separate `Production` from `DryRun` and remove `ExecutionKind` and
`ExecutionType`.
- Refactor code down
- Remove `thread_block_transaction` concept because YAGNI
- Remove `PartialBlockComponent` type
- Remove intermediate `inner` functions
- Separate `block_storage_tx` from `BlockExecutor`. This was for a
couple of reasons, but the big one is that all the `&mut self`s of the
`BlockExecutor` can only be handed out one at a time and that was making
the code more complicated. We could have just deconstructed `self`, but
then we wouldn't have access to any methods. This was a clean way to
separate them, and it makes more sense logically IMO. `ExecutorInstance`
still keeps them together, so maybe there is an argument so separate
them higher, but this is a good improvement in the mean time.

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: xgreenx <xgreenx9999@gmail.com>