Fixed reconnection issue in the dev cluster with AWS cluster (#1915)
Closes https://github.com/FuelLabs/fuel-core/issues/1592

This change improves and simplifies some aspects of the P2P service. It
also fixes the issue of not reconnecting to the reserved nodes when the
reserved node is restarted and got new IP.

- The change moves the reconnection handling into the `PeerReport`
behavior. It removes ping-ponging reserved peers between the primary
behavior and the `PeerReport` behavior and encapsulates the logic inside
the `PeerReport`. Also, it eliminates the timer and replaces it with the
queue of reconnections, reducing noise in logs(before, we had much more
trash errors).
- Added logs for cases when the dial fails. They are very helpful to
debug issues with connection.
- Simplified initialization of the `ConnectionTracker` and
`FuelAuthenticated`. It allows the reuse of libp2p built-in connections
builder.
- Removed the usage of the Mplex since it doesn't have a backpressure
mechanism. Now we use Yamux by default. It is breaking the change since
nodes with the old codebase can't connect to new nodes.
- Propagated `max_concurrent_streams` for request-response protocol.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog

### Before requesting review
- [x] I have reviewed the code myself

---------

Co-authored-by: Hannes Karppila <hannes.karppila@gmail.com>