Add Gas Price Updater Service (#1938)
Closes: https://github.com/FuelLabs/fuel-core/issues/1956

This is a subtask of https://github.com/FuelLabs/fuel-core/issues/1624

In this PR we add the generic service that will post an algorithm for
the providers to use. Not bothering with a _real_ algorithm, that will
be implemented later. For now, just show that the provider can get the
value generated by the service.

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [ ] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Swhttps://github.com/FuelLabs/fuel-core/issues/1956ay
compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: Hannes Karppila <2204863+Dentosal@users.noreply.github.com>
Co-authored-by: Green Baneling <XgreenX9999@gmail.com>