Implement gas price updater for service (#1972)
Part of: https://github.com/FuelLabs/fuel-core/issues/1955

The updater has two ports it owns:
- `L2BlockSource`
- `MetadataStorage`

`L2BlockSource` is for getting the latest block as they are committed
and using that to update the gas price algorithm.
`MetadataStorage` is for maintaining the state of the updater in the
case the service shuts down or needs to be restarted. For now it just
naively stores the `AlgorithmUpdaterV1`, and just the most recent
version. I don't know if we need to consider storing older versions of
it for the scenario we need to roll back the chain.

I started adding the DA Record stuff, but we're not using it so we can
YAGNI it for now, otherwise it would just be implementing a bunch of
adapters that don't do anything.

This implements the service without any live adapters for the ports.

Will follow this up with another PR that actually implements the ports.

## Checklist
- [ ] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [ ] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [ ] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: human <jamesturner@Zenobia.hsd1.wa.comcast.net>
Co-authored-by: Green Baneling <XgreenX9999@gmail.com>