Prepare the database to use different types than `Database` for atomic view (#1991)
## Change overview

The change implements the types for `AtomicView` and `HistoricalView`
traits. It is only preparation for corresponding features without actual
implementation.

In the past, we cloned the `Database` type and used it as a view without
any atomic properties. In a new implementation, `Database` is only a
view provider now, but the view itself implements all read
functionality.

We have two views. Separation for `IterableView` and `KeyValueView` is
needed because we can't iterate over the historical storage. We only
have access by key to the values. It would be nice to only have one view
that supports all functionality, but we don't need iteration
functionality for the historical view(at least for now), and it is
expensive to support(requires more storage or enormous computation
resources).

## Implementation details

`Database`, `IterableView` and `KeyValueView` all are aliases to the
`GenericDatabase`. The `GenericDatabase` implements all storage traits
from the `fuel-storage` crate, deriving implementation from
`StructuredStorage`. We need it reuse the storage trait implementation
for all three types. They only require the inner storage to implement
`KeyValueInspect` to derive implementation.

`KeyValueViewWrapper` and `IterableViewWrapper` are wrappers around
`Arc<dyn Trait>` that implement `KeyValueInspect` traits.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog

### Before requesting review
- [x] I have reviewed the code myself