Adapted the block producer to react on the outdated transactions from the TxPool (#2002)
Adapted the block producer to react to checked transactions that were
using another version of consensus parameters during validation in the
TxPool. After an upgrade of the consensus parameters of the network,
TxPool could store invalid `Checked` transactions. This change fixes
that by tracking the version that was used to validate the transactions.


## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself