Moved P2P database lookups into a separate thread (#2013)
When P2P service receives a large number of database related requests,
it struggles to process them and p2p events. Database lookup is a heavy
operation, and in the case of transactions, it requires a separate
lookup for each transaction. It creates a queue of p2p events that will
fail with timeouts.

It causes the node to stop syncing, since it can't process new blocks
from the network.

This change adds a separate thread to process database lookups to not
affect the p2p communication.

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself