chore(sealed_blocks): optimize fetching of sealed block header at a given height (#2112)
## Linked Issues/PRs
<!-- List of related issues/PRs -->
- #2023

## Description
<!-- List of detailed changes -->
We don't need to fetch the header *and* the consensus header from the
storage if either one of them is `None`. If `consensus` evaluates to
`None`, we can early return a `None`, but if the consensus header
exists, we can then fetch the block header.

Performs another optimization in instances where we fetch the
`SealedBlock` and only end up using the `entity` wrapped within it, and
not the consensus headers. Previously, both the consensus headers and
the block were fetched from the db and then returned as a `SealedBlock`,
Now, we just check for existence of a given block height in the
`SealedBlockConsensus` table, and avoid the allocation for the consensus
headers.

> note: also restricts how many blocks worth of transactions can be
requested over p2p

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>
Co-authored-by: Mårten Blankfors <nilssonmartenx@gmail.com>