Fix bug with gas price factor in V1 algorithm (#2201)
## Linked Issues/PRs
https://github.com/FuelLabs/fuel-core/issues/2167

## Description
Before, the final `da_gas_price` was calculated by the block producer,
but its chosen value wasn't communicated back to the updater
immediately. Instead the value was implicitly calculated by the updater
when it received the L2 block information. This was bad because it
destroyed information with respect to the gas price factor.

A value of 1 could actually have been 110 or 120 in the updater, but
then divided by the factor and rounded back to 1. On update, this would
be interpreted as just 100, and the incremental changes would be lost.

Since we are no long calculating the `da_gas_price` at the time of block
production, we could store the _actual_ `da_gas_price` value in the
updater and avoid the lost information.

Many tests existed to cover the previous behavior they were reconfigured
to match the newer, simpler behavior.

## Checklist
- [x] New behavior is reflected in tests

### Before requesting review
- [x] I have reviewed the code myself