Reject transactions that creates diamond dependencies. It allows simplify the algorithm and to nto waste a lot of resources on cyclies handling.
Removed subtracting of the collisions from the `can_fit_into_pool` method. Having this logic requires handling cases when collided transaction is sub transaction of the collision graph, or in the case of candidate graph in the `find_free_space` method.

Added graph itegrity check to be sure that its state it valid after all manipualtions.

Handle the case with zero inputs in stability test.

Re-wrote the `remove_node_and_dependent_sub_graph_recursion` to be a bfs algorithm instead of dfs.