DA compression (#1609)
Related https://github.com/FuelLabs/fuel-core/issues/1605. VM PR
https://github.com/FuelLabs/fuel-vm/pull/670.

This PR adds DA compression crate for Fuel blocks, performed upon block
creation. The compressed blocks are stored into the offchain database
and can be fetched using the GraphQL API.

## Note for reviewers

To keep this reasonably compact, decompression support is not included
in this PR, and will be done as a follow-up. As a result, the full data
roundtrip testing is not part of this PR. There's no proof here that
compression of full blocks is reversible.

## TODO
#### Features
- [x] Temporal registry db support
- [x] Optimize temporal registry eviction implementation
- [x] Implement TxId ↔ TxPointer lookups
- [x] Integrate with the block committer (GraphQL interface, probably)
#### Tests
- [x] compressed blocks are available from non-block-producer nodes
- [ ] e2e test for the full decompression cycle (moved to a follow-up)


## Follow-up issues
- Sync the node from L1:
https://github.com/FuelLabs/fuel-core/issues/2208
- Decompression roudntrip tests
https://github.com/FuelLabs/fuel-core/issues/2238
- Figure out which cache eviction algorithm/behavior is wanted:
https://github.com/FuelLabs/fuel-core/issues/2231
- Figure out if we need to remove the compressed blocks from the db
after a while
- Merkle roots for fraud proofs
https://github.com/FuelLabs/fuel-core/issues/2232

---------

Co-authored-by: Aaryamann Challani <43716372+rymnc@users.noreply.github.com>
Co-authored-by: Green Baneling <XgreenX9999@gmail.com>
Co-authored-by: Rafał Chabowski <88321181+rafal-ch@users.noreply.github.com>