Prepare the GraphQL service for the switching to `async` methods (#2334)
To avoid blocking all threads with heavy sync requests to the database,
we decided to move the main database lookups to a separate thread pool.

After we introduced `ReadView` in the GraphQL service, we duplicated all
ports for this type to avoid breaking many places in the codebase(plus
simplifying the review process of the change that caused the creation of
`ReadView`). But after adding the `ReadView` type, we don't need ports
on this type anymore since it is an internal database wrapper type of
the GraphQL service, and it has full control over how it works. So, it
was an old tech debt to remove the usage of ports. The time is come. To
simplify the introduction of the `async` and not affect ports with it,
this PR decouples ports and `ReadView`.

This PR prepares the switching from `sync` to `async` logic(it will be a
follow-up PR).

### Before requesting review
- [x] I have reviewed the code myself