Fix reverse iterator in RocksDB (#2398)
## Linked Issues/PRs
Closes https://github.com/FuelLabs/fuel-core/issues/2044

## Description

### Problem analysis

Our database columns use the prefix extractor configuration which allow
rocksdb to make optimizations on storage based on prefix. However, this
breaks the lexicographic order between prefixs and we need to pass a
special option to the read iterator to bypass the "prefix sharding"
(source:
https://github.com/facebook/rocksdb/wiki/Prefix-Seek/3a5e28faf6c2d0ec1bdb99763043e1e3322007e9#how-to-ignore-prefix-bloom-filters-in-read).

### Current solution

I took the same approach as before and started iteration from the next
prefix using the bypass read option argument.

However I change the behavior to make only iterator to simplify and
optimize the solution.

### Better solution for the future

I think that breaking the "prefix sharding" from RocksDB to iterate can
make the iterator way more costly.

There is a way to avoid it because we can iterate in reverse order
inside a prefix (source:
https://github.com/facebook/rocksdb/wiki/SeekForPrev) but we need to get
the maximum value for a key in a given prefix to use it as a starting
point of our iteration which is not possible to get in the current code
because we don't have any information about the key.

Happy to here your thoughts @FuelLabs/client :)

EDIT: Not done : https://github.com/FuelLabs/fuel-core/issues/2405

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

---------

Co-authored-by: green <xgreenx9999@gmail.com>