Update bhsh bench (#2456)
## Linked Issues/PRs
https://github.com/FuelLabs/fuel-core/issues/2237

## Description
This PR is on top of `add_zk_opcodes` so that the change of gas cost
will be efective with the new `GasCostsValueV5`. The benchmark was
incorrect because of
https://github.com/FuelLabs/fuel-core/blob/afbfef20fb78921cc6dc365878fb5b3a790ee769/crates/storage/src/vm_storage.rs#L291
. In benches the block_height was `0`. If we use the `DBBench` it works
because the block height used is 1. However, if we want to add blocks in
the db we also need the `DBBench` to use an higher block height. This PR
solves all of this.
This PR also simplify and comment the preparation code.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: Green Baneling <XgreenX9999@gmail.com>