feat(txpool|p2p): use seqlock instead of small copy-able RwLocks (#2524)
## Linked Issues/PRs
<!-- List of related issues/PRs -->
- none

## Description
<!-- List of detailed changes -->
We utilize [SeqLock](https://docs.kernel.org/locking/seqlock.html)
instead of an RwLock for data that satisfies the following conditions -

1. Occasional writes
2. Frequent reads

The `write` method no longer has a hot loop, and therefore writes aren't
blocked by reads. we don't use an internal locking mechanism, but rather
a sequence counter to check if data was changed before/during/after a
read. This may result in a few iterations while reading, but it is
minimal.

We should have a benchmark for txpool to see the performance difference.

It is important to note that this implementation *assumes* that there
will only be 1 writer thread.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: AurelienFT <32803821+AurelienFT@users.noreply.github.com>