Create new index for tracking Asset metadata (#2445)
**Note:**
Please merge https://github.com/FuelLabs/fuel-core/pull/2463 first

---

## Linked Issues/PRs
- Closes #2425

- Related to https://github.com/FuelLabs/fuels-ts/issues/3373

- It's based on top of https://github.com/FuelLabs/fuel-core/pull/2463
in order to have access to several mechanisms (like `IndexationError`)
implemented previously for other indexes.

## Description
<!-- List of detailed changes -->

This introduces a new GraphQL endpoint that returns info about an asset.
Similarly to other indexes, the asset metadata index is only available
when the client was started with no database and was able to sync from
scratch. If the index is not available, this endpoint will return an
error:
```
{
  "data": null,
  "errors": [
    {
      "message": "Asset metadata index is not available",
      "locations": [...],
    }
  ]
}
```

The off-chain worker indexes `Mint` and `Burn` events to store this
metadata about an `AssetId`.

## Checklist
- [x] New behavior is reflected in tests


### Before requesting review
- [x] I have reviewed the code myself

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [TS SDK](https://github.com/FuelLabs/fuels-ts/)

---------

Co-authored-by: Rafał Chabowski <rafal.chabowski@fuel.sh>
Co-authored-by: Green Baneling <XgreenX9999@gmail.com>
Co-authored-by: AurelienFT <aurelien.foucault@epitech.eu>
Co-authored-by: AurelienFT <32803821+AurelienFT@users.noreply.github.com>