chore(gas_price_service_v1): strictly ensure last_recorded_height is set, to avoid initial poll of da source (#2523)
> [!WARNING]
> Needs to be merged into master after #2469 is merged

## Linked Issues/PRs
<!-- List of related issues/PRs -->
- #2469 

## Description
<!-- List of detailed changes -->
When starting a node, we request the da source service for the latest
costs. This results in large negative profit / loss influencing the da
gas price. We want to avoid this behaviour, and therefore, we ensure
that the last_recorded_height is set in the block committer.

## Checklist
- [x] Breaking changes are clearly marked as such in the PR description
and changelog
- [x] New behavior is reflected in tests
- [x] [The specification](https://github.com/FuelLabs/fuel-specs/)
matches the implemented behavior (link update PR if changes are needed)

### Before requesting review
- [x] I have reviewed the code myself
- [x] I have created follow-up issues caused by this PR and linked them
here

### After merging, notify other teams

[Add or remove entries as needed]

- [ ] [Rust SDK](https://github.com/FuelLabs/fuels-rs/)
- [ ] [Sway compiler](https://github.com/FuelLabs/sway/)
- [ ] [Platform
documentation](https://github.com/FuelLabs/devrel-requests/issues/new?assignees=&labels=new+request&projects=&template=NEW-REQUEST.yml&title=%5BRequest%5D%3A+)
(for out-of-organization contributors, the person merging the PR will do
this)
- [ ] Someone else?

---------

Co-authored-by: Rafał Chabowski <88321181+rafal-ch@users.noreply.github.com>
Co-authored-by: Mitchell Turner <james.mitchell.turner@gmail.com>